% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{buttonUI}
\alias{buttonUI}
\title{按纽生成生成器界面}
\usage{
buttonUI(
  tabTitle = "按纽",
  colTitles = c("一般按纽", "下载按纽", "显示区域"),
  widthRates = c(4, 8, 12),
  func_left = buttonUI_left,
  func_right = buttonUI_right,
  func_bottom = buttonUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
buttonUI()
}
