% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{dateUI}
\alias{dateUI}
\title{demo生成器界面}
\usage{
dateUI(
  tabTitle = "日期",
  colTitles = c("操作区域", "预览区域", "显示区域"),
  widthRates = c(4, 8, 12),
  func_left = dateUI_left,
  func_right = dateUI_right,
  func_bottom = dateUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
demo生成器界面
}
\examples{
dateUI()
}
