% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceDetail.R
\name{priceDetailUI}
\alias{priceDetailUI}
\title{按纽生成生成器界面}
\usage{
priceDetailUI(
  tabTitle = "DM核价明细表",
  colTitles = c("上传数据", "显示区域"),
  widthRates = c(4, 8),
  func_left = priceDetail_left,
  func_right = priceDetail_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
priceDetailUI()
}
