% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceSummary.R
\name{priceSummaryUI}
\alias{priceSummaryUI}
\title{按纽生成生成器界面}
\usage{
priceSummaryUI(
  tabTitle = "DM核价汇总表",
  colTitles = c("上传数据", "显示区域"),
  widthRates = c(4, 8),
  func_left = priceSummary_left,
  func_right = priceSummary_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
priceSummaryUI()
}
