% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{textUI}
\alias{textUI}
\title{demo生成器界面}
\usage{
textUI(
  tabTitle = "文本框",
  colTitles = c("一般文本", "多行文本", "显示区域"),
  widthRates = c(4, 8, 12),
  func_left = textUI_left,
  func_right = textUI_right,
  func_bottom = textUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
demo生成器界面
}
\examples{
textUI()
}
